Now shows the volume in decimal (dB)

+++ Code/StatusWindow.h
@@ -2,5 +2,5 @@
 
 namespace StatusWindow {
 	extern void showBrightness(int brightnessLevel);
-	extern void showVolume(int volumeLevel);
+	extern void showVolume(double volumeLevel);
 }
